ref(tests): Conver test from enzyme to rtl and ts
